- Changed from float to double for better precision
git-svn-id: http://svn.projects.ez.no/ezgmaplocation/trunk@43 a5a2e000-4f5a-4d95-8ad8-c7c64b196f82

+++ packages/ezgmaplocation_extension/ezextension/ezgmaplocation/classes/ezgmaplocation.php
@@ -63,11 +63,11 @@ class eZGmapLocation extends eZPersistentObject
                                                         'required' => true ),
                                          'latitude' => array( 
                                                         'name' => 'latitude',
-                                                        'datatype' => 'float',
+                                                        'datatype' => 'double',
                                                         'default' => 0,
                                                         'required' => true ),
                                          'longitude' => array( 'name' => "longitude",
-                                                        'datatype' => 'float',
+                                                        'datatype' => 'double',
                                                         'default' => 0,
                                                         'required' => true ),
                                          'address' => array( 'name' => 'street',
